ninja: Support base dir beginning with ../
A base dir beginning with `../` (e.g. due to a dependency project
that exists *below* the top directory) will cause the object files
to be written to something like `$(generator_output)/Debug/obj/../x.o`
which can cause troubles with builds due to the names of object files
and output files conflicting with each other.